🔨 Changed the order of + and - in the xaml
The buttons were out of order, in the GUI, tabbing
from the PasswordLengthBox would skip over '-' and
go straight to '+', where if you tabbed again you
would go back to '-'.

This was wrong, the order was corrected to be
logical.

Tabbing from the PasswordLengthBox now brings you
to '-', where if you tab again, it brings you to
'+'.