🐛 Fixed errors in processing user input
The 3 functions that handle getting the user requested length for the
password and ensuring that it is within min and max could be made to
bypass the check if the user typed 0 or a number greater than the
maximium.

Now each of the functions check for all edge cases before accepting
the user input, this way, the bug is gone.

Additionaly the maximium number of characters was changed from 10240
to 10000 because that number is easier for humans to remember.